breaking rest.py into Routes/ and renamed to server.py
