+++ components/core/Routes/index.py
@@ -1,0 +1,2 @@
+from Routes.Download import download_blueprint
+from Routes.User import user_blueprint
